refactor: Move newNotificationAtom to store directory for better organization
